Cut nb_sim data by calendar date

Cut nb_sim data by calendar date

Usage

cut_data_by_date(data, cut_date, event_gap = 5/365.25, ...)

Arguments

data

Data generated by nb_sim().

cut_date

Calendar time (relative to trial start) at which to censor follow-up.

event_gap

Gap duration after each event during which no new events are counted. Can be a numeric value (default 5 / 365.25) or a function returning a numeric value. The time at risk is reduced by the sum of these gaps (truncated by the cut date).

...

Additional arguments passed to methods.

Value

A data frame with one row per subject randomized prior to cut_date containing the truncated follow-up time (tte) and total number of observed events (events).

Examples

enroll_rate <- data.frame(rate = 20 / (5 / 12), duration = 5 / 12)
fail_rate <- data.frame(treatment = c("Control", "Experimental"), rate = c(0.5, 0.3))
dropout_rate <- data.frame(
  treatment = c("Control", "Experimental"),
  rate = c(0.1, 0.05), duration = c(100, 100)
)
sim <- nb_sim(enroll_rate, fail_rate, dropout_rate, max_followup = 2, n = 20)
cut_data_by_date(sim, cut_date = 1)
#>    id    treatment enroll_time       tte events
#> 1   1 Experimental  0.04814947 0.9518505      0
#> 2   2      Control  0.04968518 0.8567431      0
#> 3   3 Experimental  0.08092913 0.9179037      1
#> 4   4      Control  0.10135757 0.3342787      1
#> 5   5 Experimental  0.12509265 0.8749074      0
#> 6   6 Experimental  0.12654618 0.8734538      0
#> 7   7      Control  0.13208227 0.8542285      1
#> 8   8      Control  0.13254571 0.8674543      0
#> 9   9 Experimental  0.13558272 0.8644173      0
#> 10 10 Experimental  0.19885878 0.8011412      0
#> 11 11      Control  0.20239567 0.7839151      1
#> 12 12      Control  0.22002267 0.7799773      0
#> 13 13 Experimental  0.22429347 0.7757065      0
#> 14 14      Control  0.22803670 0.7719633      0
#> 15 15 Experimental  0.23883497 0.7611650      0
#> 16 16      Control  0.25283811 0.7471619      0
#> 17 17 Experimental  0.26378892 0.7362111      0
#> 18 18 Experimental  0.29216864 0.6005228      1
#> 19 19      Control  0.32322892 0.6767711      0
#> 20 20      Control  0.37266560 0.6273344      0